Sapra/day night cycle (#25)
* Loaded the project

* Created The Day/Night Cycle

The Day/Night cycle is done. An event every hour is called with a parameter being the hour, the same with seconds, and minutes. The Numbers on all the UI is updated as such (when the event is called). You can also access the time via calling the function getHours, getMinutes, getSeconds. There's also a variable called wholeDayTime wich lets you know on a factor of 1 how much of the day has passed. And I added a rough Idea for the routines. For now, u can create a scriptable Object that has a list of time sets and positions. so, setting an element to 19:40:21 makes the object move to the position set on that time. That should be merged in the future with pathfinding. Also, right now u can only move an NPC, but it shouldn't be hard to implement different actions on each time set. Those get updated dynamically with the events set.

The time class is just a static class. To work, it's needed that the update Clock script is somewhere (now its in the canvas that I created to show the time) that just updates the clock
And I don't think I'm missing anything
You can change the scale of time to any relation of minutes to day being 3600 minutes one day real time, or setting it to 15 being 15 minutes one day
That's on that static class too

* Small changes, nothing important

* Applied changes with what @croconut said